Move SearchResult component directory location
