+++ src/views/MallSearch.vue
@@ -13,7 +13,7 @@
 import { Component, Vue } from "vue-property-decorator";
 import SearchLayout from "../components/layouts/SearchLayout.vue";
 import SearchController from "../components/search/SearchController.vue";
-import SearchResult from "../components/search/SearchResult.vue";
+import SearchResult from "../components/search/result/SearchResult.vue";
 import { Product } from "../types/model";
 
 @Component({
